Fix "storage" event handler for clear events.
This handles session clear events according to the spec as provided on
the MDN website here:
https://developer.mozilla.org/en-US/docs/Web/API/StorageEvent#Attributes

When a clear event is fired, all cached values for the given storage are
evicted.

This also resolves a NPE that would occur when a clear event was fired
from attempting to determine if the triggering key (null) matched the
managed-key pattern.